requirements.txt
bleach==6.1.0
build==1.2.2.post1
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
idna==3.10
importlib_metadata==8.5.0
kaggle==1.6.17
kagglehub==0.2.9
packaging==24.2
pip-tools==7.4.1
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
requests==2.32.3
six==1.16.0
text-unidecode==1.3
tomli==2.1.0
tqdm==4.67.0
urllib3==2.2.3
webencodings==0.5.1
zipp==3.20.2
